+++ src/main/java/ua/edu/ratos/service/session/Evaluator.java
@@ -1,12 +1,13 @@
 package ua.edu.ratos.service.session;
 
-import ua.edu.ratos.domain.entity.question.Question;
 import ua.edu.ratos.service.dto.response.*;
 
 public interface Evaluator {
-    int evaluate(ResponseMultipleChoice response);
-    int evaluate(ResponseFillBlankSingle response);
-    int evaluate(ResponseFillBlankMultiple response);
-    int evaluate(ResponseMatcher response);
-    int evaluate(ResponseSequence response);
+    double evaluate(ResponseMultipleChoice response);
+    double evaluate(ResponseFillBlankSingle response);
+    // Partial response!
+    double evaluate(ResponseFillBlankMultiple response);
+    // Partial response!
+    double evaluate(ResponseMatcher response);
+    double evaluate(ResponseSequence response);
 }
